Update README.md
I was working around your code and entered a random username and password because it was not mentioned clearly which username and password should we enter, so I thought it might be better to update it here as well. By the way, your code is beautiful and is working so far for me, I am just testing it. Thank you